pyproject.toml
[tool.poetry]
name = "unifi-mqtt"
version = "0.1.0"
description = ""
authors = ["David Cramer <dcramer@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^1.5.0"
aiohttp = "^3.6.2"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[tool.poetry.scripts]
unifi-mqtt = 'unifi_mqtt.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
